Changed templates and routes into separate files and folders. Fixes #49 and #50
