+++ openapprentice/__init__.py
@@ -83,4 +83,7 @@ user_db = peewee.MySQLDatabase(
 
 babel = Babel(application)
 
-from openapprentice.routes import main
+from openapprentice.routes.admin import admin
+from openapprentice.routes.email import email
+from openapprentice.routes.main import language, main
+from openapprentice.routes.user import login, register, manager
