+++ openapprentice/templates/admin/admin.html
