+++ openapprentice/templates/admin/admin_new_user.html
@@ -1,6 +1,6 @@
-{% extends "admin.html" %}
+{% extends "admin/admin.html" %}
 {% block body %}
-    {% from "_formhelpers.html" import render_field %}
+    {% from "utils/_formhelpers.html" import render_field %}
     <form method=post id="new-user-form">
         <div class="new-user">
             <div class="new-user-screen">
