+++ openapprentice/templates/admin/projects.html
@@ -1,7 +1,7 @@
-{% extends "admin.html" %}
+{% extends "admin/admin.html" %}
 
 {% block body %}
     <div class="col-sm-9">
-        Stats
+        Projects
     </div>
 {% endblock %}
