+++ openapprentice/templates/admin/stats.html
@@ -1,7 +1,7 @@
-{% extends "admin.html" %}
+{% extends "admin/admin.html" %}
 
 {% block body %}
     <div class="col-sm-9">
-        Projects
+        Stats
     </div>
 {% endblock %}
