+++ openapprentice/templates/admin/user_list.html
@@ -1,5 +1,5 @@
-{% extends "admin.html" %}
-{% from "user_modal_macro.html" import user_modal %}
+{% extends "admin/admin.html" %}
+{% from "admin/user_modal_macro.html" import user_modal %}
 {% block body %}
     <style>
             /* Style the buttons that are used to open and close the accordion panel */
