+++ openapprentice/templates/email/generic.html
