+++ openapprentice/templates/email/unconfirmed.html
@@ -1,4 +1,4 @@
-{% extends "main.html" %}
+{% extends "main/main.html" %}
 
 {% block body %}
