--- openapprentice/templates/flashes.html
@@ -1,13 +1,0 @@
-<!doctype html>
-{% with messages = get_flashed_messages(with_categories=true) %}
-  {% if messages %}
-    <ul class=flashes id="page-wrapper">
-    {% for category, message in messages %}
-      <li class="{{ category }}">{{ message }}</li>
-    {% endfor %}
-    </ul>
-  {% endif %}
-{% endwith %}
-{% block body %}{% endblock %}
-{% block navbar %}{% endblock %}
-{% block Footer %}{% endblock %}
