+++ openapprentice/templates/main/home.html
@@ -1,4 +1,4 @@
-{% extends "main.html" %}
+{% extends "main/main.html" %}
 {% block body %}
     <p>
         This is the home page
