+++ openapprentice/templates/main/recruiters.html
@@ -1,4 +1,4 @@
-{% extends "main.html" %}
+{% extends "main/main.html" %}
 {% block body %}
     <p>
         Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce quis ex ligula. Donec aliquet orci sed auctor lacinia. Vestibulum condimentum, purus vel dignissim dignissim, diam purus ornare sem, ac sollicitudin nibh augue vel est. Cras convallis tincidunt nisl vel ultricies. Maecenas eget elit lacinia, rhoncus dui eget, pulvinar magna. Aliquam ut sodales dolor, nec vehicula ligula. Nullam lacus ante, porttitor eget eros ut, convallis tempor massa. Nunc non condimentum massa, at facilisis lorem. Nam dignissim, lacus id pulvinar vestibulum, diam lorem dictum nunc, a efficitur ipsum velit nec sapien. Nam pretium tempor dignissim. Ut luctus mi ac tellus ornare blandit. Proin at felis vel enim sollicitudin blandit. Aliquam ac mollis odio, tempus aliquam lacus. Integer placerat maximus dui et porttitor. Vestibulum pellentesque nibh nulla, vitae feugiat tortor varius quis. Integer diam nibh, sodales at urna sit amet, ultricies cursus enim.
